* Added star search minigame to final scene * Stars light up if they are valid * Removed functionality that starts game over * TODO Connect valid stars w/ line
