+++ Assets/ClickableStar.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 4bd733d38320c8b4a92587ef8df9b026
+timeCreated: 1530052613
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
