+++ Assets/Scripts/Navigation/BeginStarSearch.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 24f1b6d12b8ee4d29a7293ec010ecd7c
+timeCreated: 1525280739
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
